add books module

+++ model/book.go
@@ -3,7 +3,7 @@ package model
 type Book struct {
 	ID      string  `json:"id"`
 	Title   string  `json:"title"`
-	Price   int     `json:"price"`
+	Price   float64 `json:"price"`
 	IsbnNo  string  `json:"isbn_no"`
 	Authors *Author `json:"authors"`
 }
